Dynamic visibility for custom appearance shapes
For circuits with custom appearance, most shapes can have a "dynamic
visibility condition" assigned to them, for example, that the shape is
only visible when a specific LED within in the circuit is on.

Currently only LEDs can provide values for the conditions, but
additional sources will be added shortly.

This partly implements feature request #93.