Fix disappearing sidebar
If the main sidebar split was adjusted quickly, it could shrink to 0 or
grow to 100%, at which point it was no longer possible to restore. E.g:
https://electronics.stackexchange.com/questions/336680/logisim-sidebar-disappeared

We now clamp the split to the range (0.05 to 0.80), which does not
prevent the problem, but at least fixes it when the window is next
opened or closed.